[tlse] tls for NovaAPI, NovaMetadata and NovNoVNCProxy
Public/Internal service cert secrets and the CA bundle secret
can be passed to configure httpd virtual hosts for tls termination.
The certs get direct mounted to the appropriate place in
/etc/pki/tls/certs/%s.crt|key and a CA bundle to
/etc/pki/ca-trust/extracted/pem/tls-ca-bundle.pem . Job deployments
for bootstrap/cron and NovaConductor, NovaScheduler and NovaCompute
get the CA bundle added if configured.

Also indexes the named input resources for CA bundle, and tls secrets
to be able to watch them for a change and reconcile.

Depends-On: https://github.com/openstack-k8s-operators/lib-common/pull/384

Jira: TODO